# expression: implement vectorized evaluation for builtinNowWithoutArgSig