@b41sh 2019-11-12
Repeat with this PR
https://github.com/pingcap/tidb/pull/13341